refactor(orders): wave-3 slice 5 order_suggestion_core scenario tables (#3949)
Migrate six order_suggestion_core_part* shards to typed scenario tables
and one thin family runner without changing suggestion semantics.
