--- packages/colonizethis_orders/test/orders/order_suggestion_core_part2_naval_mission_test.dart
@@ -1,51 +1,0 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_logic/colonizethis_logic.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_test/test.dart';
-
-void main() {
-  group('Order suggestion', () {
-    test('suggestNavalMissionOrders returns list', () {
-      const playerId = 'gp1';
-      const ow = 'oldWorld';
-      final player = const Player(
-        id: playerId,
-        displayName: 'GP',
-        isHuman: false,
-      );
-      final world = WorldState(
-        turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-        oldWorld: const RegionData(),
-        newWorld: const RegionData(),
-        fleets: [
-          Fleet(
-            id: 'fleet_gp1',
-            ownerId: playerId,
-            seaZoneId: 'sea1',
-            regionId: ow,
-            shipTypeIds: ['fluyte'],
-          ),
-        ],
-      );
-      final game = Game(id: 'g1', worldState: world, players: [player]);
-      final topology = MapTopology(
-        nodes: [
-          TopologyNode(
-            id: 'sea1',
-            regionId: ow,
-            type: TopologyNodeType.seaZone,
-          ),
-        ],
-        edges: const [],
-      );
-      final view = buildPlayerView(game, topology, playerId);
-      final suggestions = suggestNavalMissionOrders(
-        view,
-        game,
-        topology,
-        const Orders(),
-      );
-      expect(suggestions, isA<List<NavalMissionOrder>>());
-    });
-  });
-}
